puppeteer/requirements.txt
fastapi
uvicorn
httpx
cryptography
python-dotenv
apscheduler>=3.10,<4.0
pydantic
asyncpg
sqlalchemy
passlib
bcrypt<4.0.0
python-jose[cryptography]
python-multipart
psutil
psycopg2-binary
slowapi
websockets
packaging
PyJWT[crypto]>=2.7.0
pip-audit
pip-tools>=7.0
docker>=7.0.0
networkx>=3.6,<4.0
alembic==1.13